requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
drf-extra-fields==3.1.1
python-dotenv==0.18.0
psycopg2==2.9.1
Faker==8.8.1
factory_boy==3.2.0
